+++ LM15SGFNZ07.h
@@ -33,8 +33,8 @@ class LM15SGFNZ07 {
     LM15SGFNZ07(byte pinCs, byte pinReset, byte pinRs);
     void init(void);
     void clear(unsigned int color);
-    void drawBitmap(unsigned char x,unsigned char y, unsigned char width, unsigned char height, unsigned int *bitmap);
-    void drawBitmap(unsigned char x,unsigned char y, unsigned char width, unsigned char height, const PROGMEM unsigned int *bitmap);
+    void drawBitmap(unsigned char x,unsigned char y, unsigned char width, unsigned char height, unsigned short *bitmap);
+    void drawBitmap(unsigned char x,unsigned char y, unsigned char width, unsigned char height, const PROGMEM unsigned short *bitmap);
     void drawLine(unsigned char x1, unsigned char y1, unsigned x2, unsigned char y2, unsigned int color);
     void drawPixel(unsigned char x, unsigned char y,unsigned int color);
     void drawRect(unsigned char x,unsigned char y, unsigned char width, unsigned char height, unsigned int color);
